fix error
if Git is installed in MSYS2, `git rev-parse --show-toplevel` returns
path like '/c/path/to/repo'. First directory is not a real one, but a
reference to drive. `normalise` converts this to '\\c\\path\\to\\repo',
and running command will prepend current drive letter, and as result
wrond path is to be searched -- 'C:\\c\\path\\to\\repo'!
Quick solution is to strip drive reference from path. Since Windows
allows '/' only as path separator and not in names, this should be
pretty straightforward. BUT! 'Git for windows' returns path with correct
reference to drive -- 'C:/path/to/repo'. So, we have to check not only
platform (I did it with System.Info.os), but the first symbol of path,
to make sure we are not corrupting path for folks using 'Git for windows'